requirements.txt
pyTelegramBotAPI
django
psycopg2
openai-whisper
torch
torchaudio
matplotlib
pydub
word2number